Remove circular dependency between event.ts and event_database.ts
