+++ glean/src/core/metrics/events_database/index.ts
@@ -2,27 +2,23 @@
 //  * License, v. 2.0. If a copy of the MPL was not distributed with this
 //  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-import type Store from "../storage/index.js";
-import type { JSONArray, JSONObject, JSONValue } from "../utils.js";
-import type { StorageBuilder } from "../../platform/index.js";
-import { isUndefined } from "../utils.js";
-import EventMetricType from "./types/event.js";
-import log, { LoggingLevel } from "../log.js";
-import CounterMetricType from "./types/counter.js";
-import { Lifetime } from "./lifetime.js";
-import { Context } from "../context.js";
-import { generateReservedMetricIdentifiers } from "./database.js";
-
-const LOG_TAG = "core.Metric.EventsDatabase";
-
-// glean_parser will reject extra keys with `#` in the name,
-// so there extra_keys are guaranteed not to clash with user defined extra keys.
-export const GLEAN_EXECUTION_COUNTER_EXTRA_KEY = "#glean_execution_counter";
-const GLEAN_STARTUP_DATE_EXTRA_KEY = "#glean_startup_date";
-const GLEAN_RESERVED_EXTRA_KEYS = [
+import type Store from "../../storage/index.js";
+import type { JSONArray, JSONObject, JSONValue } from "../../utils.js";
+import type { StorageBuilder } from "../../../platform/index.js";
+import { isUndefined } from "../../utils.js";
+import EventMetricType from "../types/event.js";
+import log, { LoggingLevel } from "../../log.js";
+import CounterMetricType from "../types/counter.js";
+import { Lifetime } from "../lifetime.js";
+import { Context } from "../../context.js";
+import { generateReservedMetricIdentifiers } from "../database.js";
+import { RecordedEvent } from "./recorded_event.js";
+import {
   GLEAN_EXECUTION_COUNTER_EXTRA_KEY,
   GLEAN_STARTUP_DATE_EXTRA_KEY
-];
+} from "../../constants.js";
+
+const LOG_TAG = "core.Metric.EventsDatabase";
 
 /**
  * Attempts to create a date object from a string.
@@ -70,86 +66,6 @@ function getGleanRestartedEvent(sendInPings: string[]): EventMetricType {
     lifetime: Lifetime.Ping,
     disabled: false
   }, [ GLEAN_STARTUP_DATE_EXTRA_KEY ]);
-}
-
-// An helper type for the 'extra' map.
-export type ExtraMap = Record<string, string>;
-
-// Represents the recorded data for a single event.
-export class RecordedEvent {
-  constructor(
-    // The event's category.
-    //
-    // This is defined by users in the metrics file.
-    readonly category: string,
-    // The event's name.
-    //
-    // This is defined by users in the metrics file.
-    readonly name: string,
-    // The timestamp of when the event was recorded.
-    //
-    // This allows to order events.
-    readonly timestamp: number,
-    // A map of all extra data values.
-    //
-    // The set of allowed extra keys is defined by users in the metrics file.
-    public extra?: ExtraMap,
-  ) {}
-
-  static toJSONObject(e: RecordedEvent): JSONObject {
-    return {
-      "category": e.category,
-      "name": e.name,
-      "timestamp": e.timestamp,
-      "extra": e.extra,
-    };
-  }
-
-  static fromJSONObject(e: JSONObject): RecordedEvent {
-    return new RecordedEvent(
-      e["category"] as string,
-      e["name"] as string,
-      e["timestamp"] as number,
-      e["extra"] as ExtraMap | undefined
-    );
-  }
-
-  /**
-   * Add another extra key to a RecordedEvent object.
-   *
-   * @param key The key to add.
-   * @param value The value of the key.
-   */
-  addExtra(key: string, value: string): void {
-    if (!this.extra) {
-      this.extra = {};
-    }
-
-    this.extra[key] = value;
-  }
-
-  /**
-   * Generate a new RecordedEvent object,
-   * stripped of Glean reserved extra keys.
-   *
-   * @returns A new RecordedEvent object.
-   */
-  withoutReservedExtras(): RecordedEvent {
-    const extras = this.extra || {};
-    const filteredExtras = Object.keys(extras)
-      .filter(key => !GLEAN_RESERVED_EXTRA_KEYS.includes(key))
-      .reduce((obj: ExtraMap, key) => {
-        obj[key] = extras[key];
-        return obj;
-      }, {});
-
-    return new RecordedEvent(
-      this.category,
-      this.name,
-      this.timestamp,
-      (filteredExtras && Object.keys(filteredExtras).length > 0) ? filteredExtras : undefined
-    );
-  }
 }
 
 /**
