+++ glean/tests/unit/core/metrics/events_database.spec.ts
@@ -6,13 +6,15 @@ import assert from "assert";
 import Glean from "../../../../src/core/glean";
 
 import { Lifetime } from "../../../../src/core/metrics/lifetime";
-import EventsDatabase, { GLEAN_EXECUTION_COUNTER_EXTRA_KEY, RecordedEvent } from "../../../../src/core/metrics/events_database";
+import EventsDatabase from "../../../../src/core/metrics/events_database";
 import EventMetricType from "../../../../src/core/metrics/types/event";
 import type { JSONObject } from "../../../../src/core/utils";
 import CounterMetricType from "../../../../src/core/metrics/types/counter";
 import { generateReservedMetricIdentifiers } from "../../../../src/core/metrics/database";
 import PingType from "../../../../src/core/pings/ping_type";
 import { Context } from "../../../../src/core/context";
+import { RecordedEvent } from "../../../../src/core/metrics/events_database/recorded_event";
+import { GLEAN_EXECUTION_COUNTER_EXTRA_KEY } from "../../../../src/core/constants";
 
 describe("EventsDatabase", function() {
   const testAppId = `gleanjs.test.${this.title}`;
